requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
Pillow==9.5.0
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
streamlit==1.23.1
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
